Refactor to use servo in addition to 4 relays
For an implementation of this project, a 12VDC winch will be used that
will pull up to 100A, so to simplify the wiring needs, a simple servo
will be used to toggle a DPDT switch in parallel to the winch remote.
This will be a safety and cost-savings measure. The code still has 4
channel relay capabilities. Right now, channel 1 is momentary on type,
while channels 2-4 are toggle-type.